requests.cpp & requestdefines.cpp are split. requests_common & supernode_common static libraries are created.
