+++ include/lib/graft/requests.h
@@ -1,0 +1,10 @@
+
+#pragma once
+
+#include "lib/graft/router.h"
+
+namespace graft::request {
+
+void registerHealthcheckRequests(graft::Router &router);
+
+}
