+++ include/supernode/requests.h
@@ -8,7 +8,6 @@ namespace graft::supernode::request {
 void registerRTARequests(graft::Router &router);
 void registerWalletApiRequests(graft::Router &router);
 void registerForwardRequests(graft::Router &router);
-void registerHealthcheckRequests(graft::Router &router);
 void registerDebugRequests(Router &router);
 
 }
