+++ src/lib/graft/requests.cpp
@@ -1,0 +1,13 @@
+
+#include "lib/graft/requests.h"
+#include "lib/graft/requests/health_check.h"
+
+namespace graft::request {
+
+void registerHealthcheckRequests(Router &router)
+{
+    registerHealthcheckRequest(router);
+}
+
+}
+
