+++ test/cryptonode_handlers_test.cpp
@@ -33,7 +33,7 @@
 #include "lib/graft/jsonrpc.h"
 #include "lib/graft/connection.h"
 #include "lib/graft/router.h"
-#include "supernode/requests.h"
+#include "lib/graft/requests.h"
 #include "supernode/requests/get_info.h"
 #include "supernode/requests/send_raw_tx.h"
 #include "supernode/requests/authorize_rta_tx.h"
